Revert "TM-4894 redacted employee EP"
